Remove old SurfaceView
